Add error barchart with new approach
Instead of a generic plotter, I am switching to more specific pages.
While less powerful, it's going to be a lot quicker to convert the
notebook visualizations to the visualization app this way.
It will also be easier to provide context to the users of the app.

The big drawback is that it is a lot less powerful, users can't
interact a lot. Not at all currently- but I do plan to add some
interactivity, and allow users to upload their own results too.